# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Indian Options Paper Trading API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trades:
    get:
      operationId: listTrades
      tags: [trades]
      summary: List all paper trades
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [open, closed, all]
        - name: strategy
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of trades
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trade"

    post:
      operationId: createTrade
      tags: [trades]
      summary: Create a manual paper trade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTradeBody"
      responses:
        "201":
          description: Trade created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
  /trades/{id}:
    get:
      operationId: getTrade
      tags: [trades]
      summary: Get a single trade with all legs
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trade details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
        "404":
          description: Not found

    patch:
      operationId: updateTrade
      tags: [trades]
      summary: Update a trade (e.g. close it at market)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTradeBody"
      responses:
        "200":
          description: Updated trade
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
        "404":
          description: Not found

    delete:
      operationId: deleteTrade
      tags: [trades]
      summary: Delete a trade
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /trades/{id}/close:
    post:
      operationId: closeTrade
      tags: [trades]
      summary: Close a trade at current market price
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Closed trade with realized P&L
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
...
components:
  schemas:
...
    TradeLeg:
      type: object
      properties:
        id:
          type: integer
        tradeId:
          type: integer
        symbol:
          type: string
        optionType:
          type: string
          enum: [CE, PE]
        strike:
          type: number
        expiry:
          type: string
        action:
          type: string
          enum: [BUY, SELL]
        quantity:
          type: integer
        entryPrice:
          type: number
        exitPrice:
          type: ["number", "null"]
        currentPrice:
          type: number
        lotSize:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - tradeId
        - symbol
        - optionType
        - strike
        - expiry
        - action
        - quantity
        - entryPrice
        - currentPrice
        - lotSize
        - createdAt
    Trade:
      type: object
      properties:
        id:
          type: integer
        strategyType:
          type: string
          enum: [IRON_CONDOR, CALENDAR_SPREAD, INTRADAY_EXPIRY, MANUAL]
        strategyFrequency:
          type: ["string", "null"]
          enum: [WEEKLY, BIWEEKLY, MONTHLY, null]
        underlying:
          type: string
          enum: [NIFTY, BANKNIFTY, FINNIFTY]
        status:
          type: string
          enum: [open, closed]
        entryTime:
          type: string
        exitTime:
          type: ["string", "null"]
        entryUnderlyingPrice:
          type: number
        exitUnderlyingPrice:
          type: ["number", "null"]
        unrealizedPnl:
          type: number
        realizedPnl:
          type: ["number", "null"]
        maxProfit:
          type: ["number", "null"]
        maxLoss:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        legs:
          type: array
          items:
            $ref: "#/components/schemas/TradeLeg"
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - strategyType
        - underlying
        - status
        - entryTime
        - entryUnderlyingPrice
        - unrealizedPnl
        - legs
        - createdAt
        - updatedAt
    CreateTradeBody:
      type: object
      properties:
        strategyType:
          type: string
          enum: [IRON_CONDOR, CALENDAR_SPREAD, INTRADAY_EXPIRY, MANUAL]
        strategyFrequency:
          type: ["string", "null"]
          enum: [WEEKLY, BIWEEKLY, MONTHLY, null]
        underlying:
          type: string
          enum: [NIFTY, BANKNIFTY, FINNIFTY]
        notes:
          type: ["string", "null"]
        legs:
          type: array
          items:
            $ref: "#/components/schemas/CreateTradeLegBody"
      required:
        - strategyType
        - underlying
        - legs
    CreateTradeLegBody:
      type: object
      properties:
        symbol:
          type: string
        optionType:
          type: string
          enum: [CE, PE]
        strike:
          type: number
        expiry:
          type: string
        action:
          type: string
          enum: [BUY, SELL]
        quantity:
          type: integer
        entryPrice:
          type: number
        lotSize:
          type: integer
      required:
        - symbol
        - optionType
        - strike
        - expiry
        - action
        - quantity
        - entryPrice
        - lotSize
    UpdateTradeBody:
      type: object
      properties:
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [open, closed]
...