# lib/api-zod/src/generated/types/updateTradeBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Indian Options Paper Trading API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateTradeBodyStatus } from "./updateTradeBodyStatus";

export interface UpdateTradeBody {
  /** @nullable */
  notes?: string | null;
  status?: UpdateTradeBodyStatus;
}